SpaceApp/DjangoBackend/requirements.txt
django
kafka-python
djangorestframework
